# openapi.yaml
openapi: 3.1.0
info:
  title: Redubber API
  description: '# Redubber API - AI-Powered Video Redubbing Platform


    The Redubber API provides comprehensive video redubbing capabilities powered by
    OpenAI''s

    Whisper, GPT-4, and TTS technologies. Transform videos into different languages
    while

    preserving the original speaker''s tone, emotion, and delivery style.


    ## Key Features


    - **Intelligent Voice Refinement**: AI-powered voice instruction generation for
    natural-sounding TTS

    - **Multi-Voice Testing**: Generate previews for all available voices simultaneously

    - **Smart Caching**: Avoid redundant TTS generations with intelligent caching

    - **Async Processing**: High-performance async TTS with 5x throughput improvement

    - **Project Management**: Organize videos, track progress, and manage redubbing
    workflows


    ## Workflow


    1. Create a project and upload videos

    2. Start redubbing task (transcribe, translate, generate audio)

    3. Use voice refinement to test different voices and instructions

    4. Select the best voice and apply to all segments

    5. Download the final redubbed video


    ## Rate Limits


    - OpenAI API limits apply to transcription, translation, and TTS operations

    - Recommended: 1-3 concurrent redubbing tasks per instance

    - TTS generation: Up to 100 concurrent requests (configurable)'
  contact:
    name: Redubber Development Team
...
    email: support@redubber.io
  license:
...
paths:
...
  /api/tasks/redub:
    post:
      tags:
      - tasks
      summary: Submit Redub Task
      description: "Submit a video redubbing job to the task queue.\n\nValidates the\
        \ video file exists and is accessible, then enqueues\na redubbing task for\
        \ asynchronous processing. Returns immediately\nwith a task ID for status\
        \ polling.\n\nArgs:\n    task: Task creation request with video path and project\
        \ ID.\n    task_manager: TaskQueueManager injected from app state.\n\nReturns:\n\
        \    Dictionary with task_id and initial status \"queued\".\n\nRaises:\n \
        \   HTTPException: 400 if video path is invalid.\n    HTTPException: 503 if\
        \ task queue is full."
      operationId: submit_redub_task_api_tasks_redub_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
        required: true
      responses:
        '202':
          description: Successful Response
          content:
            application/json:
              schema:
                additionalProperties:
                  type: string
                type: object
                title: Response Submit Redub Task Api Tasks Redub Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tasks/{task_id}:
    get:
      tags:
      - tasks
      summary: Get Task Status
      description: "Get current status of a redubbing task.\n\nFrontend should poll\
        \ this endpoint every 2-5 seconds to track\ntask progress. Returns detailed\
        \ stage information and progress\npercentage.\n\nArgs:\n    task_id: Unique\
        \ task identifier returned from submission.\n    task_manager: TaskQueueManager\
        \ injected from app state.\n\nReturns:\n    Complete task status with progress\
        \ and current stage.\n\nRaises:\n    HTTPException: 404 if task_id not found."
      operationId: get_task_status_api_tasks__task_id__get
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskStatusResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tasks/{task_id}/cancel:
    post:
      tags:
      - tasks
      summary: Cancel Task
      description: "Cancel a running or queued redubbing task.\n\nAttempts graceful\
        \ cancellation of the task. Queued tasks are\nremoved immediately. Running\
        \ tasks complete their current stage\nbefore terminating.\n\nArgs:\n    task_id:\
        \ Task identifier to cancel.\n    task_manager: TaskQueueManager injected\
        \ from app state.\n\nReturns:\n    Cancellation confirmation with updated\
        \ status.\n\nRaises:\n    HTTPException: 404 if task_id not found.\n    HTTPException:\
        \ 409 if task is already completed or failed."
      operationId: cancel_task_api_tasks__task_id__cancel_post
      parameters:
      - name: task_id
        in: path
        required: true
        schema:
          type: string
          title: Task Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TaskCancelResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/tasks/:
    get:
      tags:
      - tasks
      summary: List Tasks
      description: "List all tasks with optional filtering.\n\nUseful for displaying\
        \ task history or monitoring all active jobs.\nResults are ordered by creation\
        \ time, most recent first.\n\nArgs:\n    status_filter: Optional filter by\
        \ status (queued/running/completed/failed).\n    project_id: Optional filter\
        \ by project ID.\n    task_manager: TaskQueueManager injected from app state.\n\
        \nReturns:\n    List of matching tasks."
      operationId: list_tasks_api_tasks__get
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - name: project_id
        in: query
        required: false
        schema:
          anyOf:
          - type: integer
          - type: 'null'
          title: Project Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskStatusResponse'
                title: Response List Tasks Api Tasks  Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    TaskCancelResponse:
      properties:
        task_id:
          type: string
          title: Task Id
          description: Cancelled task identifier
        status:
          type: string
          title: Status
          description: Updated task status after cancellation
        message:
          type: string
          title: Message
          description: Human-readable confirmation message
      type: object
      required:
      - task_id
      - status
      - message
      title: TaskCancelResponse
      description: Response schema for task cancellation.
    TaskCreate:
      properties:
        video_path:
          type: string
          title: Video Path
          description: Absolute path to video file
        project_id:
          type: integer
          title: Project Id
          description: Associated project ID
      type: object
      required:
      - video_path
      - project_id
      title: TaskCreate
      description: Request schema for submitting a redubbing task.
    TaskStatusResponse:
      properties:
        task_id:
          type: string
          title: Task Id
          description: Unique task identifier (UUID)
        video_path:
          type: string
          title: Video Path
          description: Target video file path
        status:
          type: string
          enum:
          - queued
          - running
          - completed
          - failed
          title: Status
          description: Current task status
        stage:
          type: string
          title: Stage
          description: Current processing stage description
        progress:
          type: integer
          maximum: 100.0
          minimum: 0.0
          title: Progress
          description: Progress percentage (0-100)
        error:
          anyOf:
          - type: string
          - type: 'null'
          title: Error
          description: Error message if status is 'failed'
        created_at:
          type: string
          format: date-time
          title: Created At
          description: Task creation timestamp
      type: object
      required:
      - task_id
      - video_path
      - status
      - stage
      - progress
      - created_at
      title: TaskStatusResponse
      description: Response schema for task status polling.
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...